requirements.txt
attrs==19.3.0
awsebcli==3.17.1
bcrypt==3.1.7
blessed==1.17.4
botocore==1.14.17
cached-property==1.5.1
cement==2.8.2
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.3.9
cryptography==2.9
dash==1.10.0
dash-bootstrap-components==0.9.1
dash-core-components==1.9.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
DateTime==4.3
Deprecated==1.2.7
docker==4.2.0
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
EB==0.1.5
Flask==1.1.1
Flask-Compress==1.4.0
future==0.16.0
idna==2.7
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
numpy==1.18.2
pandas==1.0.3
paramiko==2.7.1
pathspec==0.5.9
plotly==4.6.0
pycparser==2.20
PyGithub==1.47
PyJWT==1.7.1
PyNaCl==1.3.0
pyrsistent==0.16.0
python-dateutil==2.8.0
pytz==2019.3
PyYAML==5.2
requests==2.20.1
retrying==1.3.3
scipy==1.4.1
semantic-version==2.5.0
six==1.11.0
termcolor==1.1.0
texttable==1.6.2
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
Werkzeug==1.0.1
wrapt==1.12.1
zipp==3.1.0
zope.interface==5.0.2
